backend/severless_func/requirements.txt
functions-framework==3.*
google-cloud-storage==2.14.0
geopy==2.4.1